[RAM-5] Add PersonalStorageBehavior in place of player storage. Saveable
