+++ Assets/Source/Serialization/Data/PlayerData.cs
@@ -7,6 +7,7 @@ namespace Source.Serialization.Data
     public class PlayerData
     {
         public int Id;
+        public LineStorageData PersonalStorage;
         public List<ProcessorData> Processors;
         public LineStorageData MemoryStorage;
         public LineStorageData DiskStorage;
